Add files via upload
In the Movies website, the "Movies.htm" page utilizes the "get" method, 
directing actions to the "Admission.php" script. Within the "Admission.php" 
script, there is a provision for verifying the existence of the Movies class. 
The movie class is responsible for calculating the cinema ticket cost according to the age of 
the moviegoer. A base ticket price of R 50 is considered and the age is assigned to a private 
data member.